Доработка

+++ app/src/main/java/com/example/inventorycontroll/inventoryDatabase/entities/Barcode.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "barcodes")
 data class Barcode(
     @PrimaryKey(autoGenerate = true)
-    val id: Int,
-    val goodId: Int,
+    val id: Long,
+    val goodId: Long,
     val code: String
 )
